+++ sql/dbupdate.php
@@ -8,5 +8,9 @@
 ?>
 <#3>
 <?php
+//
+?>
+<#4>
+<?php
 \srag\Plugins\SrUserEnrolment\Repository::getInstance()->installTables();
 ?>